+++ beacon-chain/p2p/BUILD.bazel
@@ -19,7 +19,7 @@ go_library(
         "log.go",
         "monitoring.go",
         "options.go",
-        "pubsub_message_id.go",
+        "pubsub.go",
         "rpc_topic_mappings.go",
         "sender.go",
         "service.go",
